+++ mycodo/mycodo_flask/utils/utils_pid.py
@@ -72,7 +72,7 @@ def pid_mod(form_mod_pid_base,
         raise_output_type = Output.query.filter(
             Output.unique_id == form_mod_pid_base.raise_output_id.data).first().output_type
         if mod_pid.raise_output_id == form_mod_pid_base.raise_output_id.data:
-            if raise_output_type == 'pwm':
+            if raise_output_type in ['pwm', 'command_pwm']:
                 if not form_mod_pid_pwm_raise.validate():
                     error.append(gettext("Error in form field(s)"))
                     flash_form_errors(form_mod_pid_pwm_raise)
@@ -88,7 +88,7 @@ def pid_mod(form_mod_pid_base,
                     mod_pid.raise_max_duration = form_mod_pid_output_raise.raise_max_duration.data
                     mod_pid.raise_min_off_duration = form_mod_pid_output_raise.raise_min_off_duration.data
         else:
-            if raise_output_type == 'pwm':
+            if raise_output_type in ['pwm', 'command_pwm']:
                 mod_pid.raise_min_duration = 2
                 mod_pid.raise_max_duration = 98
             else:
@@ -103,7 +103,7 @@ def pid_mod(form_mod_pid_base,
         lower_output_type = Output.query.filter(
             Output.unique_id == form_mod_pid_base.lower_output_id.data).first().output_type
         if mod_pid.lower_output_id == form_mod_pid_base.lower_output_id.data:
-            if lower_output_type == 'pwm':
+            if lower_output_type in ['pwm', 'command_pwm']:
                 if not form_mod_pid_pwm_lower.validate():
                     error.append(gettext("Error in form field(s)"))
                     flash_form_errors(form_mod_pid_pwm_lower)
@@ -119,7 +119,7 @@ def pid_mod(form_mod_pid_base,
                     mod_pid.lower_max_duration = form_mod_pid_output_lower.lower_max_duration.data
                     mod_pid.lower_min_off_duration = form_mod_pid_output_lower.lower_min_off_duration.data
         else:
-            if lower_output_type == 'pwm':
+            if lower_output_type in ['pwm', 'command_pwm']:
                 mod_pid.lower_min_duration = 2
                 mod_pid.lower_max_duration = 98
             else: