+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fix exception-handling of sending test email ([#471](https://github.com/kizniche/mycodo/issues/471))
  - Fix HDC1000 initialization issue ([#467](https://github.com/kizniche/mycodo/issues/467))
+ - Fix Command PWM frontend issues ([#469](https://github.com/kizniche/mycodo/issues/469))
 
 
 ## 6.1.1 (2018-05-18)